refactor: reformat code from intellij
