+++ .github/workflows/tests.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v4
         with:
-          path: '.'
+          path: "."
       - name: Use Node.js 18.x
         uses: actions/setup-node@v3
         with:
