--- .prettierignore
@@ -1,12 +1,0 @@
-# Playwright Specific
-node_modules/
-test-results/
-playwright-report
-dist/
-# IDE - VSCode
-.idea
-
-# System Files
-.DS_Store
-Thumbs.db
-{"mode":"full","isActive":false}
