--- .prettierrc
@@ -1,6 +1,0 @@
-{
-  "semi": true,
-  "trailingComma": "es5",
-  "singleQuote": false,
-  "printWidth": 120
-}
