+++ playwright-tests/placeholder/api.spec.ts
@@ -1,13 +1,13 @@
-import { expect, test } from "@playwright/test";
+import {expect, test} from "@playwright/test";
 
 test.describe("API", async () => {
-  test("authentication", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("authentication", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 
-  test("unauthorized", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("unauthorized", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 });
