+++ playwright-tests/placeholder/login.spec.ts
@@ -1,13 +1,13 @@
-import { expect, test } from "@playwright/test";
+import {expect, test} from "@playwright/test";
 
 test.describe("Login", async () => {
-  test("from api", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("from api", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 
-  test("from ui", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("from ui", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 });
