+++ playwright-tests/placeholder/logout.spec.ts
@@ -1,13 +1,13 @@
-import { expect, test } from "@playwright/test";
+import {expect, test} from "@playwright/test";
 
 test.describe("Logout", async () => {
-  test("from menu", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("from menu", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 
-  test("from footer", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("from footer", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 });
