+++ playwright-tests/todo.spec.ts
@@ -1,13 +1,13 @@
-import { expect, test } from "@playwright/test";
+import {expect, test} from "@playwright/test";
 
 test.describe("TODO", async () => {
-  test("add task", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("add task", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 
-  test("delete task", async () => {
-    const status = 200;
-    expect(status).toBe(200);
-  });
+    test("delete task", async () => {
+        const status = 200;
+        expect(status).toBe(200);
+    });
 });
