+++ playwright.config.ts
@@ -1,16 +1,16 @@
-import { defineConfig } from "@playwright/test";
+import {defineConfig} from "@playwright/test";
 
 export default defineConfig({
-  testDir: "playwright-tests",
-  testMatch: "playwright-tests/**/*",
-  fullyParallel: true,
-  forbidOnly: !!process.env.CI,
-  retries: process.env.CI ? 2 : 2,
-  workers: process.env.CI ? 1 : undefined,
-  reporter: [["./src/indent-list-reporter.ts"]],
-  use: {
-    ignoreHTTPSErrors: true,
-    trace: "off",
-    colorScheme: 'dark',
-  },
+    testDir: "playwright-tests",
+    testMatch: "playwright-tests/**/*",
+    fullyParallel: true,
+    forbidOnly: !!process.env.CI,
+    retries: process.env.CI ? 2 : 2,
+    workers: process.env.CI ? 1 : undefined,
+    reporter: [["./src/indent-list-reporter.ts"]],
+    use: {
+        ignoreHTTPSErrors: true,
+        trace: "off",
+        colorScheme: "dark",
+    },
 });
